vfs0090: Use led blinking to add visual feedback on results
Once we've parsed the image results, we can now properly give a visual
feedback about the result of the previous operation.

We do this, by moving back the control to the image download SSM, where now
we parse the minutiae scanning results, and depending on it we give a visual
feedback before deactivating the device, or if needed, start a new scan.